% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_extract.R
\name{build_phrase_table}
\alias{build_phrase_table}
\title{Build phrase table of probabilities}
\usage{
build_phrase_table(model1, model2, ondisk = FALSE)
}
\description{
Takes two models (any of IBM1 through IBM4) translating in opposite directions.
For each sentence, symmetrizes the best (viterbi) alignments
and extracts phrases based on the alignments. Produces the phrase translation
probability table as an R environment object.
}
\examples{
 # data
 temp = tempfile()
 download.file("http://www.manythings.org/anki/fra-eng.zip",temp);
 ENFR = readr::read_tsv(file=unz(temp,"fra.txt"),col_names=c("en","fr","details"));
 unlink(temp);

 # a bit of pre-processing
 e = removePunctuation(ENFR$en[1:10000])
 e = str_squish(e)
 e = tolower(e)
 f = removePunctuation(ENFR$fr[1:10000])
 f = str_squish(f)
 f = tolower(f)

 # estimate models in both directions
 model1 = IBM1(target=e,source=f, maxiter=30)
 model2 = IBM1(target=f,source=e, maxiter=30)

 # phrase table (prob. of f phrase given e phrase)
 phtable = build_phrase_table(model1,model2)

 # e.g. phrase probabilities for e phrase "calm down"
 phtable[["calm down"]][["calmetoi"]]
 phtable[["calm down"]][["calmezvous"]]
 phtable[["calm down"]][["du calme"]]
 phtable[["calm down"]][["tranquille"]]

 # we can get phrase table in opposite direction by reversing model order:
 phtable2 = build_phrase_table(model2,model1)
 phtable[["calmetoi"]][["calm down"]]
 phtable[["calmetoi"]][["chill out"]]
 phtable[["calmetoi"]][["relax"]]
}
