% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CDF.R
\docType{methods}
\name{CDF}
\alias{CDF}
\title{CDF}
\usage{
CDF(trackll,dt)
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{dt}{Time intervals.}
}
\value{
\itemize{
\item{PDF} eCDF of MSD at specified dt.
}
}
\description{
Plot emperical cummulative probability (eCDF) for
  trajecotries.
}
\examples{
folder=system.file("extdata",package="smt")
trackll=readDiatrack(folder)
CDF(trackll,dt=8)
}

