% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareFolder.R
\docType{methods}
\name{compareFolder}
\alias{compareFolder}
\title{compareFolder}
\usage{
compareFolder(folder1,folder2,folder3=NULL,folder4=NULL,folder5=NULL)
}
\arguments{
\item{folder}{the path to the folder location.}
}
\value{
\itemize{
\item{trackll} A list of tracks, each item of a track list correspond to a folder. This list can then be fed into other functions for comparison.
}
}
\description{
compare (2 ≤ n ≤ 5) folders with Diatrack output files. merge track files in each folder into one item of a track list. This list can then be fed into other functions for comparison.
}
\examples{
folder1=system.file("extdata",package="smt")
folder2=system.file("extdata",package="smt")
trackll=compareFolder(folder1,folder2)
}

