% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dcoef.R
\docType{methods}
\name{Dcoef}
\alias{Dcoef}
\title{Dcoef}
\usage{
Dcoef( trackll,dt=6,filter=c(min=7,max=Inf),resolution=0.107,binwidth=0.5,method=c("static","rolling.window","percentage"),plot=c("none","histogram","density","variance"),output=F,t.interval=0.01)
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{dt}{Time intervals.}

\item{filter}{An logical indicate if frames less than specified time interval (< = dt) should be filtered out (i.e. Take only trajectories that have number of frames > dt).}

\item{resolution}{ratio of pixel to µM.}

\item{binwidth}{binwidth used for histogram, default 0.5.}

\item{plot}{A parameter for plotting. "none" (default), no plot; "histogram", plots histogram with count information, binwidth can be set through parameter binwidt; "density", plots density/frequency; "variance", plots mean and standard deviation of all trajectories, in this mode, rolling window calculaton of Dcoef is applied and filter is on.}

\item{output}{An Logical indicate if output should be generated. See Values
for detail.}

\item{t.interval}{time interval between frames, default 0.010 s (10ms).}

\item{lag.start}{time lag used as start of dt for compute Dcoef.}

\item{lag.end}{Time lag used as end of dt for compute Dcoef.}
}
\value{
\itemize{
\item{Dcoef} A list of Dcoef for each file in trackll.
\item{PDF} Log.Dcoef histogram fitted with density curve, when plot = TRUE.
\item{csv} Dcoef output in csv format, when output = TRUE.
}
}
\description{
Caclulate instantaneous diffusion coefficient (Dcoef) for
  trajecotries.
}
\details{
Generic parameters (parameter applied to all methods, such as resolution etc) are set in the function. Method dependent parameters (such as lag.start, lag.end for method = "static"), are stored in profile.csv in PREF folder under extdata. To change preference parameter, can either programably or manually go to folder system.file("extdata","PREF","profile.csv",package="smt"), and change the profile.csv.
}
\examples{
# compare files
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder)
Dcoef(trackll,plot="density")

# compare folders
folder1=system.file("extdata","SWR1",package="smt")
folder2=system.file("extdata","HTZ1",package="smt")
trackll=compareFolder(folder1,folder2)
Dcoef(trackll,plot="variance"
}

