% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_IBM1.R
\name{decode.IBM1}
\alias{decode.IBM1}
\title{(IBM1) Stack decoder}
\usage{
\method{decode}{IBM1}(...)
}
\description{
Uses the same algorithm as IBM2, and simply calls \code{decode.IBM2} with argument \code{IBM1=TRUE}.
Only difference is that in IBM1 all alignments
are equally likely. See \code{?decode.IBM2} for syntax.
}
