% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_IBM2.R
\name{decode.IBM2}
\alias{decode.IBM2}
\title{(IBM2) Compute translation probabilities given a foreign sentence.}
\usage{
\method{decode}{IBM2}(object, fsentence, max_swap = 2, numiter = 4, verbose = TRUE)
}
\arguments{
\item{object}{result from IBM2()}

\item{fsentence}{sentence in f language you'd like to translate to e language (vector or space-delimited string)}

\item{max_swap}{maximum distance away two words can be to consider swapping them during decoding (default 2).}

\item{numiter}{number of times heuristics should be tried (default=4). Usually only a couple iterations are needed to find local optimum.}

\item{verbose}{If TRUE (default), then print results of each heuristic step to the screen.}
}
\value{
The best translation in e language found after numiter rounds of heuristics.
}
\description{
Takes a sentence in a foreign language f and produces the best translation
it can find based on IBM2 model. It starts with a word-by-word translation
into e language. Then it iterates between 4 heuristics: dropping a word,
reordering words, trying alternate translations of words, or adding
alternate translations of words. These heuristics are necessary to make
decoding computationally reasonable, but may result in a local optimum
rather than a global optimum.
}
\examples{
# download english-french sentence pairs
temp = tempfile()
download.file("http://www.manythings.org/anki/fra-eng.zip",temp);
ENFR = readr::read_tsv(file=unz(temp,"fra.txt"),col_names=c("en","fr","details"));
unlink(temp);

# a bit of pre-processing
e = tolower(stringr::str_squish(tm::removePunctuation(ENFR$en[1:200])));
f = tolower(stringr::str_squish(tm::removePunctuation(ENFR$fr[1:200])));

# estimate model
out = IBM2(e,f,maxiter=50,eps=0.01,init.IBM1=5);

# possible english translations and their probabilities
besttranslation = decode(out, fsentence="une bière sil vous plaît")

}
