% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Generic decoding function.}
\usage{
decode(object, ...)
}
\description{
This generic function translates ("decodes") a given sentence in one language
to a sentence in the output language given the model estimates from \code{object}.
Currently supports decoding based on \code{IBM1} and \code{IBM2}. See \code{?decode.IBM1}
and \code{?decode.IBM2}.
}
