% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DisplacementCDF.R
\docType{methods}
\name{displacementCDF}
\alias{displacement.track}
\alias{displacementCDF}
\title{displacementCDF}
\usage{
displacementCDF(trackll,dt=6,resolution=0.107,plot=F,output=F)
}
\arguments{
\item{dt}{Time intervals.}

\item{resolution}{ratio of pixel to µM.}

\item{trackll}{Track list output from readDiatrack().}

\item{plot}{An logical indicate if plot should be generated. See Values for
detail.}

\item{output}{An logical indicate if output should be generated. See Values
for detail.}
}
\value{
\itemize{
\item{ list of "stepwise.displacement" and "CDF.displacement",} {A list of stepwise.displacement" and "CDF.displacement". the name of the list is the Diatrack folder name.}

\item{Output file,} {Displacement of individual trajectoreis at specified dt.}


\item{CDF plot,} {CDF plot of displacement for individual files.
}
}
}
\description{
calculate cumulative distribution function of all displacement for individual trajectories.
}
\examples{
folder1=system.file("extdata","SWR1",package="smt")
folder2=system.file("extdata","HTZ1",package="smt")
trackll=compareFolder(c(folder1,folder2))
displacementCDF(trackll,dt=1,plot=T)
}

