% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM2.R
\name{IBM2}
\alias{IBM2}
\title{IBM2 Model}
\usage{
IBM2(
  e,
  f,
  maxiter = 30,
  eps = 0.01,
  init.IBM1 = 10,
  add.null.token = TRUE,
  init.tmatrix = NULL,
  init.amatrix = NULL
)
}
\arguments{
\item{e}{vector of sentences in language we want to translate to}

\item{f}{vector of sentences in language we want to translate from}

\item{maxiter}{max number of EM iterations allowed}

\item{eps}{convergence criteria for perplexity (i.e. negative log-likelihood)}

\item{init.IBM1}{number of iterations (integer>=0) of IBM1 to perform to initialize IBM2 algorithm}

\item{add.null.token}{If TRUE (default), adds \if{html}{\out{<NULL>}} to beginning of each f sentence. Allows e words to be aligned with "nothing".}

\item{init.tmatrix}{tmatrix from a previous estimation. If not provided, algorithm starts with uniform probabilities.}

\item{init.amatrix}{amatrix from a previous estimation. If not provided, algorithm starts with uniform probabilities.}
}
\value{
\item{tmatrix}{Environment object containing translation probabilities for e-f word pairs. E.g. tmatrix$go$va (equivalently, tmatrix[\link{"go"}][\link{"va"}]) gives the probability of e="go" given f="va".}
\item{amatrix}{A list of alignment probability matrices. E.g. amatrix[\link{3}][\link{4}] gives the alignment probability matrix for e sentences of length 3 and f sentences of length 4.}
\item{numiter}{Number of iterations}
\item{maxiter}{As above}
\item{eps}{As above}
\item{converged}{TRUE if algorithm stopped once eps criteria met. FALSE otherwise.}
\item{perplexity}{Final likelihood/perplexity value.}
\item{time_elapsed}{Time in minutes the algorithm ran for.}
}
\description{
The second SMT model from Brown et al. (1993)
}
\examples{
# download english-french sentence pairs
temp = tempfile()
download.file("http://www.manythings.org/anki/fra-eng.zip",temp);
ENFR = readr::read_tsv(file=unz(temp,"fra.txt"),col_names=c("en","fr","details"));
unlink(temp);

# a bit of pre-processing
e = tolower(stringr::str_squish(tm::removePunctuation(ENFR$en[1:1000])));
f = tolower(stringr::str_squish(tm::removePunctuation(ENFR$fr[1:1000])));

# IBM1 guarantees global optimum, while IBM2 doesn't => initialize with IBM1.
# IBM1 also faster, may make IBM2 faster overall.
# Experiment with number of initial IBM1 iterations:
test0 = SMT::IBM2(e,f,maxiter=50,eps=0.01,init.IBM1=0);
test1 = SMT::IBM2(e,f,maxiter=50,eps=0.01,init.IBM1=1);
test2 = SMT::IBM2(e,f,maxiter=50,eps=0.01,init.IBM1=2);
test3 = SMT::IBM2(e,f,maxiter=50,eps=0.01,init.IBM1=3);
}
