% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM3.R
\name{IBM3}
\alias{IBM3}
\title{IBM3 Model}
\usage{
IBM3(
  e,
  f,
  maxiter = 30,
  eps = 0.01,
  heuristic = TRUE,
  maxfert = 5,
  init.IBM1 = 5,
  init.IBM2 = 10,
  sparse = FALSE,
  fmatch = FALSE
)
}
\arguments{
\item{e}{vector of sentences in language we want to translate to}

\item{f}{vector of sentences in language we want to translate from}

\item{maxiter}{max number of EM iterations allowed}

\item{eps}{convergence criteria for perplexity (i.e. negative log-likelihood)}

\item{heuristic}{If TRUE (default) use a heuristic hill-climbing algorithm to find most likely alignments. If FALSE, search over all alignments (not recommended unless only looking at small sentences.)}

\item{maxfert}{Maximum number of e words ("fertility") which an f word is allowed to be mapped to. The default is 5. In practice fertility tends to be small, so this number should normally be sufficient.}

\item{init.IBM1}{number of iterations (integer>=0) of IBM1 to perform to initialize IBM2 algorithm}

\item{init.IBM2}{number of iterations (integer>=0) of IBM2 to perform to initialize IBM3 algorithm}

\item{sparse}{If TRUE, uses sparse matrices from Matrix package (default FALSE).}
}
\value{
\item{tmatrix}{Matrix of translation probabilities (cols are words from e, rows are words from f). If sparse=TRUE, tmatrix will be a sparseMatrix from the Matrix package, and will generally take up substantially less memory.}
\item{dmatrix}{A list of distortion probability matrices. E.g. dmatrix[\link{3}][\link{4}] gives the distortion probability matrix for e sentences of length 3 and f sentences of length 4.}
\item{fertmatrix}{Matrix of fertility probabilities.}
\item{p_null}{Probability of NULL token insertion.}
\item{numiter}{Number of iterations}
\item{maxiter}{As above}
\item{eps}{As above}
\item{converged}{TRUE if algorithm stopped once eps criteria met. FALSE otherwise.}
\item{perplexity}{Final likelihood/perplexity value.}
\item{time_elapsed}{Time in minutes the algorithm ran for.}
}
\description{
The third SMT model from Brown et al. (1993)
}
\examples{
# download english-french sentence pairs
temp = tempfile()
download.file("http://www.manythings.org/anki/fra-eng.zip",temp);
ENFR = readr::read_tsv(file=unz(temp,"fra.txt"),col_names=c("en","fr","details"));
unlink(temp);

# a bit of pre-processing
e = tolower(stringr::str_squish(tm::removePunctuation(ENFR$en[1:200])));
f = tolower(stringr::str_squish(tm::removePunctuation(ENFR$fr[1:200])));

# a model
model3 = IBM3(e=e,f=f,maxiter=10, init.IBM1=10, init.IBM2=20)

}
