% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_extract.R
\name{phrase_extract}
\alias{phrase_extract}
\title{Extract phrases from alignment points}
\usage{
phrase_extract(align_points, l1, l2)
}
\description{
Takes a list of alignment points (e.g. as produced by symalign function)
and returns a list of phrase pairs extracted from the alignments.
Based on algorithm provided in Koehn (2009).
}
