% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTrack.R
\docType{methods}
\name{plotTrack}
\alias{.plotTrack}
\alias{plotTrack}
\title{plotTrack}
\usage{
plotTrack(ab.trackll,resolution=0.107,frame.min=8,frame.max=100,frame.start=1,frame.end=500)
}
\arguments{
\item{resolution}{ratio of pixel to µM.}

\item{frame.min}{minimum frame number for plotting.}

\item{frame.max}{max frame number for plotting.}

\item{frame.start}{the first frame to plot. Default 1.}

\item{frame.end}{last frame to plot. Default 500.}

\item{ab.trackll}{absolute coordinates for plotting, generated from readDiatrack(folder,ab.track=T).}
}
\value{
\itemize{
\item{PDF} One PDF file with all the frames satisfy the creteria. If trackll has multiple items, it ouptus mutiple PDF files each corresponding to one item.
}
}
\description{
Plot track/trajectory from track list. either randomly or specified.
}
\details{
plotTrackFromIndex: if user provide a csv file with first column listing the index of trajectories, this program will plot the tracks isted in the csv file. It is useful after manipulating with the output from Dceof, to plot the tracks that of interest to the user (e.g. highest Dcoef). User need to provide the indexFile.csv, and specify the movie folder which contains the movies where specified trajectories are tracked.
}
\examples{
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder,ab.track=TRUE)
plotTrack(trackll)

## plot from index file
index.file=system.file("extdata","INDEX","indexFile.csv",package="smt")
movie.folder=system.file("extdata","SWR1",package="smt")
plotTrackFromIndex(index.file=index.file,movie.folder = movie.folder)
}

