% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_tagger.R
\name{pos.tagger}
\alias{pos.tagger}
\title{Part-of-Speech (POS) Tagger}
\usage{
pos.tagger(sen, udpipe.model)
}
\arguments{
\item{sen}{A sentence (character string)}

\item{udpipe.model}{a pre-trained tagger model returned from a udpipe::udpipe_load_model call}
}
\value{
Return a character string representing a segmented and tagged version of sen. See example below.
}
\description{
This a simple wrapper for udpipe POS tagging. It takes a single sentence (can be longer text as well) and returns the same sentence segmented and with POS tags in square brackets attached to each word.
}
\examples{
m_eng = udpipe::udpipe_download_model(language = "english-ewt")
m_eng = udpipe::udpipe_load_model(m_eng)
pos.tagger("This is an example",m_eng) # produces "this[PRON] is[AUX] an[DET] example[NOUN]"
}
