% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_IBM1.R
\name{predict_IBM1}
\alias{predict_IBM1}
\title{(IBM1) Compute translation probabilities given a foreign sentence.}
\usage{
predict_IBM1(
  fsentence,
  tmatrix,
  threshold = 0.001,
  maxlength = length(unlist(stringr::str_split(fsentence, pattern = " ")))
)
}
\arguments{
\item{fsentence}{sentence in language you'd like to translate (vector or space-delimited string)}

\item{tmatrix}{as output from IBM1()}

\item{threshold}{reduce the number of words to consider by eliminating very unlikely ones with Prob<=threshold (default=0.001).}

\item{maxlength}{only consider translations which have maxlength words or fewer. By default, will look for translations with the same number of words (or fewer) as fsentence.}
}
\value{
A data.frame with two columns, sorted in descending order of pr:
\item{esentence}{Possible translation of fsentence.}
\item{pr}{Translation probability (up to a constant).}
}
\description{
Takes a sentence in a foreign language and produces a list of possible translations and their probabilities based on the IBM1 model.
}
\examples{
# download english-french sentence pairs
temp = tempfile()
download.file("http://www.manythings.org/anki/fra-eng.zip",temp);
ENFR = readr::read_tsv(file=unz(temp,"fra.txt"),col_names=c("en","fr","details"));
unlink(temp);

# a bit of pre-processing
e = tolower(stringr::str_squish(tm::removePunctuation(ENFR$en[1:200])));
f = tolower(stringr::str_squish(tm::removePunctuation(ENFR$fr[1:200])));

# estimate model
out = IBM1(e,f,maxiter=50,eps=0.01);

# possible english translations and their probabilities
predict.IBM1(fsentence="une bière sil vous plaît",tmatrix=out$tmatrix)
}
