% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{SMT_demo}
\alias{SMT_demo}
\title{Demo SMT}
\usage{
SMT_demo(
  num_items = 3L,
  feedback = SMT::SMT_feedback_with_graph(),
  admin_password = "demo",
  researcher_email = "longgoldstudy@gmail.com",
  dict = SMT::SMT_dict,
  language = "en",
  ...
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test. Defaults to a graph-based feedback page.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.
Defaults to \code{"demo"}.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.
Defaults to \email{longgoldstudy@gmail.com},
the email address of this package's developer.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{language}{The language you want to run your demo in.
Possible languages include English (\code{"en"}) and German (\code{"de"}..
The first language is selected by default}

\item{...}{Further arguments to be passed to \code{\link{SMT}()}.}
}
\description{
This function launches a demo for the SMT.
}
