% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{SMT_feedback_with_graph}
\alias{SMT_feedback_with_graph}
\title{SMT feedback (with graph)}
\usage{
SMT_feedback_with_graph(dict = SMT::SMT_dict, graph = "pie")
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
SMT_demo(feedback = SMT_feedback_with_score())}
}
