% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/squareDisp.R
\docType{methods}
\name{squareDisp}
\alias{squareDisp}
\title{squareDisp}
\usage{
squareDisp(folder,merge=F)
}
\arguments{
\item{folder}{Full path to Diatrack output file.}

\item{merge}{An logical indicate if the output list should be merged into one. Default merge = FALSE, output list is divided by file names.}
}
\value{
\itemize{
\item{merge = F} Defult. A list of list of data.frames. First level is a list of file names in Diatrack output folder, second level is a list of data.frames from individual output file.
\item{merge = T} A list list of data.frames. First level is the folder name. second level is a list of data.frames from all Diatrack output files merged into one
}
}
\description{
read output file (tracks/trajecotries) from Diatrack.
}
\details{
default merge = FALSE, so the researcher can assay variations between files. Keep both output as two level list is for simplicity of downstream analysis.
}
\examples{
folder=system.file("extdata","SWR1",package="smt")
trackll=squareDisp(folder)
str(trackll)
}

