% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_extract.R
\name{symalign}
\alias{symalign}
\title{Symmetrization of Word Alignments}
\usage{
symalign(a1, a2)
}
\description{
Takes two vectors representing alignments of two sentences going in opposite directions.
Uses the heuristic algorithm described in Koehn (2009) to combine the alignments
into a single set of "symmetrized" alignments.
The result is a list of length-2 vectors representing the final set of alignment points.
}
