% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_align.R
\name{viterbi_align}
\alias{viterbi_align}
\title{Viterbi alignments for IBM1 and IBM2}
\usage{
viterbi_align(object, target_sen, source_sen)
}
\description{
Finds best alignment for two sentences given IBM1 or IBM2 model estimates.
}
