% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dinst.R
\docType{methods}
\name{Dcoef}
\alias{Dcoef}
\title{Dcoef}
\usage{
Dcoef(trackll,dt=8,resolution=0.107,lag.start=2,lag.end=5,filter=F,binwidth=0.5,method=c("static","rolling.window","percentage"),plot=c("none","histogram","density","variance"),output=F)
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{dt}{Time intervals.}

\item{resolution}{ratio of pixel to µM.}

\item{lag.start}{time lag used as start of dt for compute Dcoef.}

\item{lag.end}{Time lag used as end of dt for compute Dcoef.}

\item{filter}{An logical indicate if frames less than specified time interval (< = dt) should be filtered out (i.e. Take only trajectories that have number of frames > dt).}

\item{binwidth}{binwidth used for histogram, default 0.5.}

\item{plot}{A parameter for plotting. "none" (default), no plot; "histogram", plots histogram with count information, binwidth can be set through parameter binwidt; "density", plots density/frequency; "variance", plots mean and standard deviation of all trajectories, in this mode, rolling window calculaton of Dcoef is applied and filter is on.}

\item{output}{An Logical indicate if output should be generated. See Values
for detail.}
}
\value{
\itemize{
\item{Dcoef} A list of Dcoef for each file in trackll.
\item{PDF} Log.Dcoef histogram fitted with density curve, when plot = TRUE.
\item{csv} Dcoef output in csv format, when output = TRUE.
}
}
\description{
Caclulate instantaneous diffusion coefficient (Dcoef) for
  trajecotries.
}
\examples{
# compare files
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder)
Dcoef(trackll,plot="density")

# compare folders
folder1=system.file("extdata","SWR1",package="smt")
folder2=system.file("extdata","HTZ1",package="smt")
trackll=compareFolder(folder1,folder2)
Dcoef(trackll,plot="variance")
}

