% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Generic decoding function.}
\usage{
decode(object, ...)
}
\description{
This generic function translates ("decodes") a given sentence in one language
to a sentence in the output language given the model estimates from \code{object}.
Currently supports decoding based on \code{IBM1}, \code{IBM2}, or phrase-based. See \code{?decode.IBM1},
\code{?decode.IBM2}, \code{decode.phrase_table}.
}
