% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dinst.R
\docType{methods}
\name{Dinst}
\alias{Dinst}
\title{Dinst}
\usage{
Dinst(trackll,dt=8,resolution=0.107,lag.start=2,lag.end=5,filter=F,plot=F,output=F)
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{dt}{Time intervals.}

\item{resolution}{ratio of pixel to µM.}

\item{lag.start}{time lag used as start of dt for compute Dinst.}

\item{lag.end}{Time lag used as end of dt for compute Dinst.}

\item{filter}{An logical indicate if frames less than specified time interval (< = dt) should be filtered out (i.e. Take only trajectories that have number of frames > dt).}

\item{plot}{An Logical indicate if plot should be generated. See Values for
detail.}

\item{output}{An Logical indicate if output should be generated. See Values
for detail.}
}
\value{
\itemize{
\item{Dinst} A list of Dinst for each file in trackll.
\item{PDF} Log.Dinst histogram fitted with density curve, when plot = TRUE.
\item{csv} Dinst output in csv format, when output = TRUE.
}
}
\description{
Caclulate instantaneous diffusion coefficient (Dinst) for
  trajecotries.
}
\examples{
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder)
Dinst(trackll,plot=TRUE)
}

