% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitCDF.R
\docType{methods}
\name{fitCDF}
\alias{fitCDF}
\alias{one.comp.fit}
\title{fitCDF}
\usage{
fitCDF(cdf,
       components=c("one","two","three"),
       start.value=list(
            oneCompFit=list(D=0.5),
            twoCompFit=list(D1=0.1,D2=0.5,alpha=0.5),
            threeCompFit=list(D1=0.3,D2=0.3,D3=0.3,alpha=0.3,beta=0.5)))
}
\arguments{
\item{start.value}{the start value for fitting.}

\item{cdf}{cdf calculated from displacementCDF().}

\item{components}{parameter specifying the number of components to fit.Currently support one to three components fit.}
}
\value{
\itemize{
\item{on screen output,} Result and parameters of goodness of the fit.
\item{Plot,} fiting plot.
}
}
\description{
Caclulate apparent diffusion coefficient (Dcoef) for
  trajecotries by fitting displacementCDF.
}
\details{
calculating Dceof by fitting displacementCDF.
}
\examples{
# compare folders
folder1=system.file("extdata","SWR1",package="smt")
folder2=system.file("extdata","HTZ1",package="smt")
trackll=compareFolder(c(folder1,folder2))
cdf=displacementCDF(trackll,dt=1,plot=F,output=F)
fitCDF(cdf,components="two")
}

