% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitNormDistr.R
\docType{methods}
\name{fitNormDistr}
\alias{fitNormDistr}
\title{fitNormDistr}
\usage{
fitNormDistr(dcoef,components=2,log.transform=F,output=F)
}
\arguments{
\item{dcoef}{diffusion coefficient calculated from Dcoef().}

\item{components}{parameter specifying the number of components to fit.}

\item{log.transform}{logical indicate if log10 transformation is needed, default F.}

\item{output}{Logical indicaring if output file should be generated.}
}
\value{
\describe{
\item{proportions}{The final mixing proportions.}
\item{mean}{The final mean parameters.}
\item{sd}{The final standard deviations.}
\item{loglik}{The log likelihood.}
}
}
\description{
fit normal distribution to diffusion coefficient caclulated by Dcoef method.
}
\examples{
# compare folders
folder1=system.file("extdata","SWR1",package="smt")
folder2=system.file("extdata","HTZ1",package="smt")
trackll=compareFolder(c(folder1,folder2))
dcoef=Dcoef(trackll,dt=6,plot="none",output=F)
fitNormDistr(dcoef,components=2,log.transform=F,output=F)
}

