% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msd.R
\docType{methods}
\name{msd}
\alias{msd}
\alias{msd.track}
\title{msd}
\usage{
msd(trackll,dt=6,resolution=0.107,summarize=F,filter=c(min=7,max=Inf),plot=F,output=F)
}
\arguments{
\item{dt}{Time intervals.}

\item{resolution}{ratio of pixel to µM.}

\item{trackll}{Track list output from readDiatrack().}

\item{summarize}{An logical indicate if MSD should be calculated on
  individual trajectories (Default) or summarized on all trajectories.}

\item{filter}{a vector specifies the minimum and max length of trajecotries to be analyzed. Take only trajectories that has number of frames greater than (>=) min and less than (<) max.}

\item{plot}{An logical indicate if plot should be generated. See Values for
detail.}

\item{output}{An logical indicate if output should be generated. See Values
for detail.}
}
\value{
\itemize{
\item{SummarizedMSD} MSD summarized over all trajectories as a function of
dt. \item{InidvidualMSD} MSD of individual trajectories at specified dt. Row
number corresponding to its dt. Notice only the trajectories that satisfies
the specified dt is output, trajectories that does not satisfy (i.e.
trajectories satisfies 1:(dt-1)) is not output here.

\item{StandardError} Standard Error of the sample mean measures the
variations of sample mean to underlying mean, it is estimated as
SE=SD/sqrt(N).


\item{SampleSize} The sample size (number of tracks/trajectories) used for
calculating the msd and standard error.
}
}
\description{
calculate mean square displacement for individual trajectory or
  summarize on trajectories.
}
\details{
msd() calculate track (/trajectory)'s mean square displacement as a function
of time (dt). For a track of N steps, at each dt, there are dt number of
sub-trajectory/sub-tracks, mean of dt-wise sub-trajectories/ step-wise sub
tracks average subtracks into one number at each dt.

the dt number of su-btracks each contains N:N-dt steps. Because minimum step
is 1 (N-dt > = 1), so the maxium dt is N-1 (dt < = N-1). As dt increase, the
number of steps used to generate that mean decrease with the maxmum dt
(dt=N-1) generated from one step.

if one wants to focus on a group of trajectory's evolution, he can simply filter on a number that is bigger than the dt he wanted to plot MSD.
}
\examples{
folder=system.file("extdata","SWR1",package="smt")
trackll=readDiatrack(folder)
msd=msd(trackll,dt=8,summarize=TRUE,plot=TRUE)
str(msd)

## focus on a group of trajectory by setting filter greater than dt
msd=msd(trackll,dt=6,summarize=TRUE,filter=c(7,Inf),plot=TRUE)
}

