% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTrack.R
\docType{methods}
\name{plotTrack}
\alias{.plotTrack}
\alias{plotTrack}
\title{plotTrack}
\usage{
plotTrack(trackll,exposure=10,plot=F,output=F)
}
\arguments{
\item{trackll}{Track list output from readDiatrack().}

\item{exposure}{Exposure time, default = 10ms.}

\item{plot}{An Logical indicate if plot should be generated. If plot = TRUE, the plot data will also be output.}

\item{output}{An Logical indicate if output should be generated. 1) dwell time of tracks in the track list output to csv file. Each item in the list will have an individual csv file. 2) Plot PDF and plot data will be saved.}
}
\value{
\itemize{
\item{dwell time list} A list of dwell time for every trajectory, separated by file names of the trajectory file in Diatrack file folder. If combined dewell time is intended, use readDiatrack(folder, merge=T) to generate a single length list, then apply this function.
\item{PDF} dwell time frequency plot in PDF format, when plot = TRUE.
\item{csv} dwell time output in csv format, when output = TRUE.
}
}
\description{
Plot track/trajectory from track list. either randomly or specified.
}
\examples{
folder=system.file("extdata",package="smt")
trackll=readDiatrack(folder)
plotTrack(trackll,plot=T)
}

