% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readDiatrack.R
\docType{methods}
\name{readDiatrack}
\alias{readDiatrack}
\title{readDiatrack}
\usage{
readDiatrack(folder,divide=T)
}
\arguments{
\item{folder}{Full path to Diatrack output file.}

\item{divide}{An logical indicate if the output list should be divided by file name. Default divide = TRUE.}
}
\value{
\itemize{
\item{divide = T} Defult. A list of list of data.frame. First level is a list of file names in Diatrack output folder, second level is a list of data.frames from individual output file.
\item{divide = F} A list of data.frames from all Diatrack output files.
}
}
\description{
read output file (tracks/trajecotries) from Diatrack.
}
\details{
possibility of a reads been random
possibility of peak heights (base coverage) at specific location is esitmated from the average (lambda) coverage of the peak region, 1kb region, 5kb region and 10 kb regions.
This is to account for local fluctuations.
}
\examples{
folder=system.file("extdata",package="smt")
trackll=readDiatrack(folder)
str(trackll)
trackl=readDiatrack(folder,divide = F)
str(trackl)
}

