% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{SMT_feedback_with_score}
\alias{SMT_feedback_with_score}
\title{SMT feedback (with score)}
\usage{
SMT_feedback_with_score(dict = SMT::SMT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
\examples{
\dontrun{
SMT_demo(feedback = SMT_feedback_with_score())}
}
